Fixes #27 - Moved Resources and ResourcePanel components out into their own file like they're damn well supposed to be.
